database: Share `DBMigrator` code between apis as well
