+++ audio-api/src/main/scala/no/ndla/audioapi/MainClass.scala
@@ -28,7 +28,7 @@ class MainClass(val props: AudioApiProperties) extends NdlaTapirMain {
   def beforeStart(): Unit = {
     logger.info("Starting DB Migration")
     val dBstartMillis = System.currentTimeMillis()
-    componentRegistry.migrator.migrate(): Unit
+    componentRegistry.migrator.migrate()
     logger.info(s"Done DB Migration took ${System.currentTimeMillis() - dBstartMillis} ms")
   }
