+++ draft-api/src/main/scala/no/ndla/draftapi/MainClass.scala
@@ -27,7 +27,7 @@ class MainClass(override val props: DraftApiProperties) extends NdlaTapirMain {
   override def beforeStart(): Unit = {
     logger.info("Starting the db migration...")
     val startDBMillis = System.currentTimeMillis()
-    componentRegistry.migrator.migrate(): Unit
+    componentRegistry.migrator.migrate()
     logger.info(s"Done db migration, took ${System.currentTimeMillis() - startDBMillis}ms")
   }
