+++ frontpage-api/src/main/scala/no/ndla/frontpageapi/ComponentRegistry.scala
@@ -9,7 +9,7 @@ package no.ndla.frontpageapi
 
 import com.zaxxer.hikari.HikariDataSource
 import no.ndla.common.Clock
-import no.ndla.database.DataSource
+import no.ndla.database.{DBMigrator, DataSource}
 import no.ndla.frontpageapi.controller.*
 import no.ndla.frontpageapi.model.api.ErrorHandling
 import no.ndla.frontpageapi.model.domain.{DBFilmFrontPageData, DBFrontPageData, DBSubjectFrontPageData}
@@ -39,7 +39,7 @@ class ComponentRegistry(properties: FrontpageApiProperties)
     with ConverterService
     with SwaggerDocControllerConfig {
   override val props: FrontpageApiProperties = properties
-  override val migrator                      = new DBMigrator
+  override val migrator: DBMigrator          = DBMigrator()
   override val dataSource: HikariDataSource  = DataSource.getHikariDataSource
   DataSource.connectToDatabase()
