--- frontpage-api/src/main/scala/no/ndla/frontpageapi/DBMigrator.scala
@@ -1,33 +1,0 @@
-/*
- * Part of NDLA frontpage-api
- * Copyright (C) 2018 NDLA
- *
- * See LICENSE
- */
-
-package no.ndla.frontpageapi
-
-import no.ndla.database.DataSource
-import org.flywaydb.core.Flyway
-import org.flywaydb.core.api.output.MigrateResult
-
-trait DBMigrator {
-  this: Props with DataSource =>
-  val migrator: DBMigrator
-
-  class DBMigrator {
-    def migrate(): MigrateResult = {
-      val flyway = Flyway
-        .configure()
-        .javaMigrations()
-        .locations("no/ndla/frontpageapi/db/migration")
-        .dataSource(dataSource)
-        // Seems like flyway uses datasource.getConnection().getScheme() which is null if the scheme does not exist.
-        // Therefore we simply override it with dataSource.getScheme.
-        // https://github.com/flyway/flyway/issues/2182
-        .schemas(dataSource.getSchema)
-        .load()
-      flyway.migrate()
-    }
-  }
-}
