+++ frontpage-api/src/test/scala/no/ndla/frontpageapi/TestEnvironment.scala
@@ -9,7 +9,7 @@ package no.ndla.frontpageapi
 
 import com.zaxxer.hikari.HikariDataSource
 import no.ndla.common.Clock
-import no.ndla.database.DataSource
+import no.ndla.database.{DBMigrator, DataSource}
 import no.ndla.frontpageapi.controller.{FilmPageController, FrontPageController, SubjectPageController}
 import no.ndla.frontpageapi.model.api.ErrorHandling
 import no.ndla.frontpageapi.model.domain.{DBFilmFrontPageData, DBFrontPageData, DBSubjectFrontPageData}
