+++ learningpath-api/src/main/scala/no/ndla/learningpathapi/MainClass.scala
@@ -28,7 +28,7 @@ class MainClass(override val props: LearningpathApiProperties) extends NdlaTapir
   override def beforeStart(): Unit = {
     logger.info("Starting the db migration...")
     val startDBMillis = System.currentTimeMillis()
-    componentRegistry.migrator.migrate(): Unit
+    componentRegistry.migrator.migrate()
     logger.info(s"Done db migration, took ${System.currentTimeMillis() - startDBMillis}ms")
   }
